postular90%

+++ app/Http/Controllers/postulante/PostulanteController.php
@@ -146,12 +146,12 @@ class PostulanteController extends Controller
         
             return $query;
         }
-    public function experiencias_data1(){
+    public function experiencias_data1(Request $data){
             $query = ExperienciaLabUser::where('user_id',auth()->user()->id)->orderBy('id','DESC')->get();
-            
-                return $query;
+            $proceso = Proceso::select('consid_prac_preprof','consid_prac_prof')->where('id',$data->idproceso)->get();
+                return compact('query','proceso'); //REVISAR CÓDIGO DESPUÉS
             }
-
+/*
     public function formacion_data(){
 
         $query = FormacionUser::where('user_id','2')->select('grado_id','especialidad','centro_estudios','fecha_expedicion')->get();
@@ -173,7 +173,7 @@ class PostulanteController extends Controller
         }
         return json_encode($data, true);  
 
-}
+} */
 
     public function guardarformacion(Request $data){
         
@@ -285,10 +285,21 @@ class PostulanteController extends Controller
         
         $Exper = ExperienciaLabUser::find($data->id);
         $Exper->delete();
-        
-        return "eliminadoEXPERIENCIA";
+        $suma_expgen = ExperienciaLabUser::select('dias_exp_gen')
+        ->where('user_id',auth()->user()->id)
+        ->sum('dias_exp_gen');
+        $suma_expesp = ExperienciaLabUser::select('dias_exp_esp')
+        ->where('user_id',auth()->user()->id)
+        ->sum('dias_exp_esp');
+
+        return compact('suma_expgen','suma_expesp');
+            
     }
     
+    public function editarformacion(Request $data){
+        $query = FormacionUser::where('id',$data->id)->get();
+        return $query;
+    }
     public function editarexperiencia(Request $data){
         $query = ExperienciaLabUser::where('id',$data->id)->get();
         return $query;
@@ -334,7 +345,7 @@ class PostulanteController extends Controller
         $suma_expesp = ExperienciaLabUser::select('dias_exp_esp')
         ->where('user_id',auth()->user()->id)
         ->sum('dias_exp_esp');
-        $proceso = Proceso::where('id',10)->get();
+        $proceso = Proceso::where('id',$data->idproceso)->get();
         $min_expgen = $proceso[0]->anios_exp_lab_gen;
         $min_expesp = $proceso[0]->anios_exp_lab_esp;
